+++ website/static/js/loadRqbUtils.js
@@ -10,6 +10,8 @@ if (ExecutionEnvironment.canUseDOM) {
     globalThis.parseJsonLogic = parseJsonLogic;
     const { parseMongoDB } = await import('react-querybuilder/parseMongoDB');
     globalThis.parseMongoDB = parseMongoDB;
+    const { parseSpEL } = await import('react-querybuilder/parseSpEL');
+    globalThis.parseSpEL = parseSpEL;
     const { parseSQL } = await import('react-querybuilder/parseSQL');
     globalThis.parseSQL = parseSQL;
     const { transformQuery } = await import('react-querybuilder/transformQuery');
@@ -20,6 +22,7 @@ if (ExecutionEnvironment.canUseDOM) {
  • parseCEL
  • parseJsonLogic
  • parseMongoDB
+ • parseSpEL
  • parseSQL
  • transformQuery`
     );