+++ CHANGELOG.md
@@ -91,6 +91,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - [#537] New API documentation, generated directly from the source code, at https://react-querybuilder.js.org/api. In support of this, many types and functions now have better JSDoc comments which should provide a better developer experience in modern IDEs.
 - [#638] Value selectors now respect the `disabled` property of individual options in option lists.
 - [#663] The `format` option set during a call to `formatQuery` will be passed to custom rule processors as a property of the options object in the second parameter.
+- [#663] `parseSpEL` method for importing [SpEL](https://docs.spring.io/spring-framework/docs/3.0.x/reference/expressions.html) expressions.
 
 ### Fixed